add command line arguments like the C++ lib
