+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "rust-analyzer.cargo.features": [
+        "default",
+        "rgbmatrix-mock",
+        "args"
+    ]
+}
