+++ Cargo.toml
@@ -15,10 +15,10 @@ readme = "README.md"
 edition = "2018"
 
 [features]
-
 default = ["embeddedgraphics"]
 
 embeddedgraphics = ["embedded-graphics"]
+args = ["clap"]
 
 [build-dependencies]
 gcc = "0.3"
@@ -26,6 +26,11 @@ gcc = "0.3"
 [dependencies]
 libc = "0.2"
 embedded-graphics = { version = "0.6.2", optional = true }
+clap = { version = "2.33", optional = true }
+
+[[example]]
+name = "arguments"
+required-features = ["args"]
 
 [profile.release]
 codegen-units = 1
